Hubi porzadkuje potrzeba jeszcze napisac funkcje dla wartosci
