--- ToDo.txt
@@ -1,7 +1,0 @@
-Hubert Kuc
-Master Thesis
-Started 02.05.2017
-
-ToDo
-//copy what you have got on a paper
-
