+++ Training/Blender/Windshield.mtl
@@ -1,0 +1,13 @@
+# Blender MTL File: 'Windshield.blend'
+# Material Count: 1
+
+newmtl Glass
+Ns 249.019608
+Ka 0.500000 0.500000 0.500000
+Kd 0.000000 0.000000 0.000000
+Ks 1.000000 1.000000 1.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 7
+map_Kd textures/backwindow.png
