+++ Training/Blender/mikepan_bmw3v3.mtl
@@ -1,0 +1,287 @@
+# Blender MTL File: 'mikepan_bmw3v3.blend'
+# Material Count: 28
+
+newmtl Ambient_Control
+Ns 96.078431
+Ka 1.000000 1.000000 1.000000
+Kd 0.181148 0.181148 0.181148
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 1
+
+newmtl AngelEye
+Ns 17.647059
+Ka 2.000000 2.000000 2.000000
+Kd 0.500000 0.373313 0.250000
+Ks 0.100000 0.100000 0.100000
+Ke 1.000000 0.746625 0.500000
+Ni 1.000000
+d 1.000000
+illum 2
+
+newmtl BMWBlack
+Ns 390.196078
+Ka 0.100000 0.100000 0.100000
+Kd 0.000000 0.000000 0.000000
+Ks 0.300000 0.300000 0.300000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl BMWBlue
+Ns 390.196078
+Ka 0.100000 0.100000 0.100000
+Kd 0.047403 0.241055 0.629930
+Ks 0.300000 0.300000 0.300000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl BMWSilver
+Ns 96.078431
+Ka 0.600000 0.600000 0.600000
+Kd 0.139176 0.139176 0.139176
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl BMWWhite
+Ns 390.196078
+Ka 0.100000 0.100000 0.100000
+Kd 0.800000 0.800000 0.800000
+Ks 0.300000 0.300000 0.300000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl Body
+Ns 60.784314
+Ka 0.161811 0.161811 0.161811
+Kd 0.583562 0.583562 0.583562
+Ks 0.200000 0.200000 0.200000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl Bolt
+Ns 29.411765
+Ka 2.000000 2.000000 2.000000
+Kd 0.362296 0.362296 0.362296
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 2
+
+newmtl BreakLight
+Ns 123.529412
+Ka 0.200000 0.200000 0.200000
+Kd 0.447988 0.026375 0.017389
+Ks 1.000000 1.000000 1.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+map_Kd textures/Taillight.png
+
+newmtl BreakPad
+Ns 123.529412
+Ka 0.800000 0.800000 0.800000
+Kd 0.603827 0.603827 0.603827
+Ks 0.800000 0.800000 0.800000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 2
+map_Kd textures/wm_normal_abb.jpg
+
+newmtl Chrome
+Ns 123.529412
+Ka 1.000000 1.000000 1.000000
+Kd 0.301914 0.301914 0.301914
+Ks 0.500000 0.500000 0.500000
+Ke 0.060383 0.060383 0.060383
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl ChromeMatte
+Ns 123.529412
+Ka 0.700000 0.700000 0.700000
+Kd 0.301914 0.301914 0.301914
+Ks 1.000000 1.000000 1.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl Foglight
+Ns 231.372549
+Ka 0.500000 0.500000 0.500000
+Kd 0.000000 0.000000 0.000000
+Ks 0.200000 0.200000 0.200000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl Glass
+Ns 249.019608
+Ka 0.500000 0.500000 0.500000
+Kd 0.000000 0.000000 0.000000
+Ks 1.000000 1.000000 1.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 0.600000
+illum 7
+map_Kd textures/backwindow.png
+
+newmtl Glass.Tinted
+Ns 249.019608
+Ka 1.000000 1.000000 1.000000
+Kd 0.000000 0.000000 0.000000
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 0.900000
+illum 1
+
+newmtl HeadLightGlass
+Ns 45.098039
+Ka 1.000000 1.000000 1.000000
+Kd 0.787412 0.861383 1.000000
+Ks 1.000000 1.000000 1.000000
+Ke 0.393706 0.430692 0.500000
+Ni 1.100000
+d 0.900000
+illum 5
+
+newmtl Interior
+Ns 37.254902
+Ka 1.000000 1.000000 1.000000
+Kd 0.042404 0.042404 0.042404
+Ks 0.020000 0.020000 0.020000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 2
+
+newmtl Internal
+Ns 37.254902
+Ka 1.200000 1.200000 1.200000
+Kd 0.028612 0.028612 0.028612
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 1
+
+newmtl Material
+Ns 17.647059
+Ka 1.000000 1.000000 1.000000
+Kd 0.081994 0.081994 0.081994
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 1
+
+newmtl Material.001
+Ns 996.078431
+Ka 0.100000 0.100000 0.100000
+Kd 0.640000 0.000000 0.000000
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl Material.002
+Ns 96.078431
+Ka 0.500000 0.500000 0.500000
+Kd 0.492209 0.183895 0.023825
+Ks 0.500000 0.500000 0.500000
+Ke 0.307631 0.114934 0.014891
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl Matte
+Ns 37.254902
+Ka 2.000000 2.000000 2.000000
+Kd 0.241531 0.241531 0.241531
+Ks 0.100000 0.100000 0.100000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 2
+
+newmtl MatteBlack
+Ns 37.254902
+Ka 0.051000 0.051000 0.051000
+Kd 0.042404 0.042404 0.042404
+Ks 0.050000 0.050000 0.050000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl Mirror
+Ns 194.117647
+Ka 1.000000 1.000000 1.000000
+Kd 0.000000 0.000000 0.000000
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl RIm
+Ns 29.411765
+Ka 0.400000 0.400000 0.400000
+Kd 0.301914 0.301914 0.301914
+Ks 1.000000 1.000000 1.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
+
+newmtl StudioLightPanel
+Ns 96.078431
+Ka 2.000000 2.000000 2.000000
+Kd 0.800000 0.800000 0.800000
+Ks 0.500000 0.500000 0.500000
+Ke 2.000000 2.000000 2.000000
+Ni 1.000000
+d 1.000000
+illum 2
+
+newmtl TireRubber
+Ns 17.647059
+Ka 1.000000 1.000000 1.000000
+Kd 0.114978 0.114978 0.114978
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 1
+map_Kd textures/sidewall.jpg
+map_Bump textures/sidewall.jpg
+
+newmtl grill
+Ns 68.627451
+Ka 0.500000 0.500000 0.500000
+Kd 0.301914 0.301914 0.301914
+Ks 0.695041 0.695041 0.695041
+Ke 0.000000 0.000000 0.000000
+Ni 1.000000
+d 1.000000
+illum 5
