--- skrypt.sh
@@ -1,3 +1,0 @@
-export LD_LIBRARY_PATH=~/Programowanie/OSG/OpenSceneGraph/lib
-export OSG_FILE_PATH=~/Programowanie/OSG/OpenSceneGraph-Data-3.0.0
-
