+++ windshieldImage.cpp
@@ -1,0 +1,64 @@
+#include "windshieldImage.h"
+
+
+namespace MasterThesisProject{
+	windshieldImage::windshieldImage(){
+
+
+	}
+
+	bool  windshieldImage::attachWindshieldAsTexture(){
+		std::cout <<"Hubi to mistrz" << std::endl;
+
+	}
+
+
+
+	bool windshieldImage::setPositionWindshield(){
+
+
+
+
+
+	}
+
+
+
+
+	bool windshieldImage::setLight(){
+
+
+
+
+
+
+	}
+
+
+
+	osg::ref_ptr<osg::Node> windshieldImage::getNode(){
+
+
+
+
+
+	}
+
+	void windshieldImage::setCamera(){
+
+
+
+
+
+	}
+
+	void windshieldImage::setTextureSize(){
+		 tex_width = 1024;
+		 tex_height = 1024;
+
+	}
+
+	osg::ref_ptr<osg::Camera> windshieldImage::getWindshieldCamera(){
+
+	}
+}
